toolkit/components/passwordmgr/test/test_0init.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test initialization for Password Manager</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Password Manager test: initialization.
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Password Manger: initialization **/

netscape.security.PrivilegeManager.enablePrivilege('UniversalXPConnect');


// Get the pwmgr service
var Cc_pwmgr = Components.classes["@mozilla.org/passwordmanager;1"];
ok(Cc_pwmgr != null, "Access Cc[@mozilla.org/passwordmanager;1]");

var Ci_pwmgr = Components.interfaces.nsIPasswordManager;
ok(Ci_pwmgr != null, "Access Ci.nsIPasswordManager");

var pwmgr = Cc_pwmgr.getService(Ci_pwmgr);
ok(pwmgr != null, "pwmgr getService()");


// Make sure it's not already initialized somehow
var count, enum;
count = 0;
enum = pwmgr.enumerator;
while (enum.hasMoreElements()) { count++; enum.getNext(); }
ok(count == 0, "ensure no pre-existing logins");
if (count != 0) { throw "Aborting test."; }

count = 0;
enum = pwmgr.rejectEnumerator;
while (enum.hasMoreElements()) { count++; enum.getNext(); }
ok(count == 0, "ensure no pre-existing disabled logins");
if (count != 0) { throw "Aborting test"; }


pwmgr = pwmgr.QueryInterface(Components.interfaces.nsIPasswordManagerInternal);
// Add some logins for testing.
const TESTHOST = "http://localhost:8888";
for (var u = 1; u <= 15; u++) {
  pwmgr.addUserFull2(TESTHOST, "testuser"+u, "testpass"+u, "uname"+u, "pword"+u, TESTHOST);
}

// Simple check to see if everything was added fine.
count = 0;
enum = pwmgr.enumerator;
while (enum.hasMoreElements()) { count++; enum.getNext(); }
ok(count == 15, "check number of added logins");

count = 0;
enum = pwmgr.rejectEnumerator;
while (enum.hasMoreElements()) { count++; enum.getNext(); }
ok(count == 0, "check no disable logins");

</script>
</pre>
</body>
</html>
